Cinema and User Service Complete

+++ back-end/services/cinema/src/application/usecases/deleteCine/delete_cine.ts
@@ -2,13 +2,21 @@ import { Cinema } from "@prisma/client";
 import { prisma } from "../../../prisma/client";
 import { deleteCineDTO } from "../../../dtos/delete_cine_dto";
 
-export class deleteUser{
-    async execute({id}: deleteCineDTO): Promise<Cinema> {
-        const deleteCine = await prisma.cinema.delete({
+export class deleteCine{
+    async execute({id}: deleteCineDTO){
+      const cineVerify = await prisma.cinema.findUnique({
+        where: {
+            id
+        }
+     });
+      if(!cineVerify){
+        return 400;
+        }
+        await prisma.cinema.delete({
             where: {
               id
             },
           })
-        return deleteCine;
+        return 200;
     }
 }
